making sure there is a coupon code set first
It's possible that a user visits a PMP site with the url set to discount_code=code in which case the code isn't actually set as it's passed through here.

This solves the immediate problem of trying to check part of an object when it is not set but it would be possible to have a code set then and NOT record the affiliate payment because the code is not set in the object.

That fix is probably a patch to PMP but for now this keeps AffWP from throwing errors. It wouldn't have been catching these codes in the first place.